third_party/blink/web_tests/svg/custom/circle-update-to-empty-expected.txt

CONSOLE ERROR: Error: <circle> attribute r: A negative value is not valid. ("-50")
